fixed try/catch structure in movie screen

+++ src/pages/api/delete-movie.ts
@@ -40,7 +40,9 @@ export default function deleteFiles(){
         })
 
         moviesToDelete.map(async (movie: MovieData) => {
-            fs.rmSync(`./movies/${movie.imdb_code}`, { recursive: true, force: true });
+            if (fs.existsSync(`./movies/${movie.imdb_code}`)) { // make sure this works properly
+                fs.rmSync(`./movies/${movie.imdb_code}`, { recursive: true, force: true });
+            }
             await prisma.movies.delete({
                 where: {
                     imdb_code: movie.imdb_code
